theory_evaluation - Copy/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope
asttokens
asyncpg==0.29.0
background==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm
debugpy
decorator
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup
executing
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jedi
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline
mdurl==0.1.2
openai==1.33.0
orjson==3.10.3
packaging
parso
pexpect
platformdirs
prompt-toolkit
psutil
psycopg==3.1.19
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess
pure-eval
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
Pygments
python-dateutil
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq
requests==2.32.3
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
six
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data 
starlette==0.37.2
tornado
traitlets
typer==0.12.3
typing_extensions
ujson==5.10.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth
websockets==12.0
wheel==0.43.0
zipp
